ai-reading-lab/requirements.txt
pydantic>=2.7
PyMuPDF>=1.24.0
tiktoken>=0.7.0
openai>=1.40.0
tenacity>=8.3.0
python-dotenv>=1.0.1
pytest>=8.2.0
ruff>=0.6.0
black>=24.8.0

# OCR dependencies
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Crawler dependencies
requests>=2.31.0
urllib3>=2.0.0

# Notification dependencies (email is built-in, requests already included)
